+++ src/lib.rs
@@ -57,7 +57,9 @@ extern "C" fn enable_mmu() {
 
 /// Sets the appropriate vbar to point to our `vector_table`, if the `exceptions` feature is
 /// enabled.
-extern "C" fn set_exception_vector() {
+///
+/// If `exceptions` is not enabled then this is a no-op.
+pub extern "C" fn set_exception_vector() {
     // SAFETY: We provide a valid vector table.
     #[cfg(all(feature = "el1", feature = "exceptions"))]
     unsafe {