+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Unreleased
+
+### Breaking changes
+
+- Added `ExceptionHandlers` trait and `exception_handlers!` macro to define exception handlers,
+  rather than relying on global symbols.
+- Added `RegisterStateRef` parameter to exception handlers to access the saved register state, and
+  removed ELR and SPSR parameters.
+
 ## 0.3.1
 
 ### Bugfixes