+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   rather than relying on global symbols.
 - Added `RegisterStateRef` parameter to exception handlers to access the saved register state, and
   removed ELR and SPSR parameters.
+- `start_core` now takes an `FnOnce` rather than an `fn` and an argument. This allows arbitrary
+  context to be captured in the closure.
 
 ## 0.3.1