+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Improvements
+
+- Made `set_exception_vector` public.
+
 ## 0.4.0
 
 ### Breaking changes