+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 
 - Made `set_exception_vector` public.
+- Made `enable_mmu` public.
 
 ## 0.4.0